# Improve sanitize_slug