Improve the `sanitize_slug` method for better performance.


## Approach

- `sanitize_slug` is used when needed only.


## QA notes

- Ensure images are delivered from Cloudinary and the admin settings work.