@yonran 2025-10-07
Thank you! I tested your branch and found that it had an equivalent performance improvement to my #1091. This PR reduces median request time (compared to master) by about 123ms (tracing disabled) and with flamegraph trace enabled (`xdebug.mode=trace`, `xdebug.trace_format=3`, `xdebug.use_compression=true`, ` --cookie XDEBUG_TRACE=1`), it reduces `Cloudinary\Plugin->setup` time from 1.5s to 221ms (saves 1.3s).

master: setup takes 1.5s with trace enabled:
<img width="2400" height="1232" alt="Screenshot 2025-10-07 at 1 56 46 PM" src="https://github.com/user-attachments/assets/aaa87d6a-e856-4f8c-a419-e64a3674aae1" />

This PR #1101: setup takes 220ms with trace enabled:
<img width="2402" height="1265" alt="Screenshot 2025-10-07 at 1 56 57 PM" src="https://github.com/user-attachments/assets/60f82f9d-1f1b-4c69-918a-053ab5dd0bed" />

My previous PR #1091: setup takes 205ms with trace enabled (negligible difference from this PR)
<img width="2399" height="1261" alt="Screenshot 2025-10-07 at 1 57 05 PM" src="https://github.com/user-attachments/assets/eb703e5f-9e09-4da8-aac0-991ef2a60c0f" />